+++ selfdrive/car/volkswagen/interface.py
@@ -115,6 +115,10 @@ class CarInterface(CarInterfaceBase):
       ret.mass = 1205 + STD_CARGO_KG
       ret.wheelbase = 2.61
 
+    elif candidate == CAR.AUDI_Q3_MK2:
+      ret.mass = 1623 + STD_CARGO_KG
+      ret.wheelbase = 2.68
+
     elif candidate == CAR.SEAT_ATECA_MK1:
       ret.mass = 1900 + STD_CARGO_KG
       ret.wheelbase = 2.64