+++ selfdrive/controls/lib/latcontrol_angle.py
@@ -1,4 +1,5 @@
 import math
+
 from cereal import log
 
 
@@ -21,5 +22,7 @@ class LatControlAngle():
       angle_steers_des += params.angleOffsetDeg
 
     angle_log.saturated = False
-    angle_log.steeringAngleDeg = angle_steers_des
+    angle_log.steeringAngleDeg = float(CS.steeringAngleDeg)
+    angle_log.steeringAngleDesiredDeg = angle_steers_des 
+
     return 0, float(angle_steers_des), angle_log