+++ selfdrive/manager/build.py
@@ -12,9 +12,9 @@ from common.spinner import Spinner
 from common.text_window import TextWindow
 from selfdrive.hardware import TICI
 from selfdrive.swaglog import cloudlog, add_file_handler
-from selfdrive.version import dirty
+from selfdrive.version import get_dirty
 
-MAX_CACHE_SIZE = 2e9
+MAX_CACHE_SIZE = 4e9 if "CI" in os.environ else 2e9
 CACHE_DIR = Path("/data/scons_cache" if TICI else "/tmp/scons_cache")
 
 TOTAL_SCONS_NODES = 2405
@@ -77,7 +77,7 @@ def build(spinner, dirty=False):
         # Show TextWindow
         spinner.close()
         if not os.getenv("CI"):
-          error_s = "\n \n".join(["\n".join(textwrap.wrap(e, 65)) for e in errors])
+          error_s = "\n \n".join("\n".join(textwrap.wrap(e, 65)) for e in errors)
           with TextWindow("openpilot failed to build\n \n" + error_s) as t:
             t.wait_for_exit()
         exit(1)
@@ -98,4 +98,4 @@ def build(spinner, dirty=False):
 if __name__ == "__main__" and not PREBUILT:
   spinner = Spinner()
   spinner.update_progress(0, 100)
-  build(spinner, dirty)
+  build(spinner, get_dirty())